[330. Patching Array][Time Limit Exceeded]committed by Krunning

+++ 330-Patching-Array/solution.java
@@ -1,6 +1,6 @@
 public class Solution {
     public int minPatches(int[] nums, int n) {
-        long miss = 1, added = 0, i = 0;
+        int miss = 1, added = 0, i = 0;
         while (miss <= n) {
             if (i < nums.length && nums[i] <= miss) {
                 miss += nums[i++];
